Refactored filtering to be part of envelope logic
