+++ Adapter/src/main/java/uk/co/aquaq/kdb/adapter/AdapterApplication.java
@@ -5,43 +5,42 @@ import uk.co.aquaq.kdb.adapter.utils.AdapterProperties;
 import uk.co.aquaq.kdb.adapter.utils.PropertyFileLoader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import java.nio.file.Files;
-import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Properties;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 public class AdapterApplication {
 
-  private static Logger log = LoggerFactory.getLogger(AdapterApplication.class);
+  private static final Logger log = LoggerFactory.getLogger(AdapterApplication.class);
   private static AtomicBoolean noStopFile = new AtomicBoolean(true);
 
   public static void main(String[] args) {
 
     try {
 
-      final PropertyFileLoader properties = new PropertyFileLoader();
-      final Properties props =
+      final var properties = new PropertyFileLoader();
+      final var props =
           properties.getPropValues(
               args.length > 0 ? args[0] : "src\\main\\resources\\application.properties");
-      final AdapterProperties adapterProperties = new AdapterProperties(props);
+      final var adapterProperties = new AdapterProperties(props);
 
-      final ChronicleToKdbAdapter adapter =
+      final var adapter =
           new ChronicleToKdbAdapter(adapterProperties.getAdapterMessageType(), adapterProperties);
 
-      // Check application.properties for runMode...
-      // adapter.runMode=NORMAL -> normal mode processing messages on queue to kdb+
-      // adapter.runMode=KDB_BENCH -> simple testing of batched (envelope) kdb+ writes only
-
-      Thread thread = new Thread(adapter);
-      thread.start();
+      ExecutorService executor = Executors.newSingleThreadExecutor();
+      Future<?> future = executor.submit(adapter);
 
       while (noStopFile.get()) {
+        if (future.isDone()) break;
         checkForStopSignal(adapterProperties, adapter);
-        TimeUnit.MILLISECONDS.sleep(adapterProperties.getStopFileCheckInterval());
+        TimeUnit.SECONDS.sleep(adapterProperties.getStopFileCheckInterval());
       }
+
+      executor.shutdownNow();
 
     } catch (Exception ex) {
       log.error("Problem running Adapter: Exception: {}", ex.toString());
@@ -51,7 +50,7 @@ public class AdapterApplication {
   public static void checkForStopSignal(
       AdapterProperties adapterProperties, ChronicleToKdbAdapter adapter) {
     try {
-      Path path = Paths.get(adapterProperties.getStopFile());
+      var path = Paths.get(adapterProperties.getStopFile());
       if (Files.exists(path)) {
         log.info("Stop file present. Stop running adapter");
         adapter.stop();
