+++ Adapter/src/main/java/uk/co/aquaq/kdb/adapter/customexceptions/AdapterConfigurationException.java
@@ -1,8 +1,12 @@
 package uk.co.aquaq.kdb.adapter.customexceptions;
 
 public class AdapterConfigurationException extends RuntimeException {
-    public AdapterConfigurationException(String message) {
-        super(message);
-    }
 
+  public AdapterConfigurationException(String message) {
+    super(message);
+  }
+
+  public AdapterConfigurationException(String message, Throwable cause) {
+    super(message, cause);
+  }
 }
