+++ Adapter/src/main/java/uk/co/aquaq/kdb/adapter/customexceptions/KdbException.java
@@ -1,8 +1,12 @@
 package uk.co.aquaq.kdb.adapter.customexceptions;
 
 public class KdbException extends RuntimeException {
-    public KdbException(String message) {
-        super(message);
-    }
 
+  public KdbException(String message) {
+    super(message);
+  }
+
+  public KdbException(String message, Throwable cause) {
+    super(message, cause);
+  }
 }
