+++ Adapter/src/main/java/uk/co/aquaq/kdb/adapter/messages/MessageTypes.java
@@ -3,8 +3,8 @@ package uk.co.aquaq.kdb.adapter.messages;
 public class MessageTypes {
 
   public enum AdapterMessageTypes {
-    QUOTE("uk.co.aquaq.kdb.adapter.messages.ChronicleQuoteMsg"),
-    TRADE("uk.co.aquaq.kdb.adapter.messages.ChronicleTradeMsg");
+    QUOTE("uk.co.aquaq.kdb.adapter.messages.KdbQuoteMessage"),
+    TRADE("uk.co.aquaq.kdb.adapter.messages.KdbTradeMessage");
 
     public final String implClass;
