+++ Adapter/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 # What mode the Adapter will run in - NORMAL, BENCH or KDB_BENCH
-adapter.runMode=NORMAL
+adapter.runMode=KDB_BENCH
 
 # Location and name of stop file. If this exists the Adapter will exit / not start
 adapter.stopFile=C:\\ChronicleQueue\\Producer\\quote\\STOP.txt
@@ -18,10 +18,10 @@ adapter.messageType=QUOTE
 # Filter message IN if Filter Field value is in this comma separated list
 #adapter.messageFilterIn=JUVE.MI,VOD.L,HEIN.AS
 # Filter message OUT if Filter Field value in this comma separated list
-#adapter.messageFilterOut=JUVE.MI
+# e.g. adapter.messageFilterOut=JUVE.MI
 
 # Redundant. May come back in.
-adapter.waitTime.whenNoMsgs=50
+#adapter.waitTime.whenNoMsgs=50
 
 # Filesystem location of Chronicle queue to process
 chronicle.source=C:\\ChronicleQueue\\Producer\\quote
